Split constraint/approx.rs
